GROK-17560: Fitting: In-worker optimization
Fixed dependencies
