--- libraries/compute-utils/function-views/src/fitting/diff-studio/constants.ts
@@ -1,1 +1,0 @@
-export const ARG_IDX = 0;
